reorg requests (#161)
* reorg requests

* optional bools -> bool

* while i'm here fix some types in keypairs

* docs

* cleanup

* comments

* cleanup after rebase

* use cryptoalgorithm where possible in models
