+++ .pre-commit-config.yaml
@@ -35,4 +35,5 @@ repos:
         entry: poetry run sphinx-apidoc -o docs/source xrpl
         language: system
         pass_filenames: false
+        files: xrpl/.*__init__
         types: [python]
