--- docs/source/xrpl.models.requests.accounts.rst
@@ -1,85 +1,0 @@
-xrpl.models.requests.accounts package
-=====================================
-
-Submodules
-----------
-
-xrpl.models.requests.accounts.account\_channels module
-------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_channels
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.account\_currencies module
---------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_currencies
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.account\_info module
---------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_info
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.account\_lines module
----------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_lines
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.account\_objects module
------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_objects
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.account\_offers module
-----------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_offers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.account\_tx module
-------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.account_tx
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.gateway\_balances module
-------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.gateway_balances
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.accounts.no\_ripple\_check module
-------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.accounts.no_ripple_check
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: xrpl.models.requests.accounts
-   :members:
-   :undoc-members:
-   :show-inheritance:
