--- docs/source/xrpl.models.requests.paths.rst
@@ -1,45 +1,0 @@
-xrpl.models.requests.paths package
-==================================
-
-Submodules
-----------
-
-xrpl.models.requests.paths.book\_offers module
-----------------------------------------------
-
-.. automodule:: xrpl.models.requests.paths.book_offers
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.paths.deposit\_authorized module
------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.paths.deposit_authorized
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.paths.path\_find module
---------------------------------------------
-
-.. automodule:: xrpl.models.requests.paths.path_find
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.paths.ripple\_path\_find module
-----------------------------------------------------
-
-.. automodule:: xrpl.models.requests.paths.ripple_path_find
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: xrpl.models.requests.paths
-   :members:
-   :undoc-members:
-   :show-inheritance:
