+++ docs/source/xrpl.models.requests.rst
@@ -1,17 +1,72 @@
 xrpl.models.requests package
 ============================
 
-Subpackages
------------
-
-.. toctree::
-   :maxdepth: 4
-
-   xrpl.models.requests.accounts
-   xrpl.models.requests.paths
-
 Submodules
 ----------
+
+xrpl.models.requests.account\_channels module
+---------------------------------------------
+
+.. automodule:: xrpl.models.requests.account_channels
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.account\_currencies module
+-----------------------------------------------
+
+.. automodule:: xrpl.models.requests.account_currencies
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.account\_info module
+-----------------------------------------
+
+.. automodule:: xrpl.models.requests.account_info
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.account\_lines module
+------------------------------------------
+
+.. automodule:: xrpl.models.requests.account_lines
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.account\_objects module
+--------------------------------------------
+
+.. automodule:: xrpl.models.requests.account_objects
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.account\_offers module
+-------------------------------------------
+
+.. automodule:: xrpl.models.requests.account_offers
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.account\_tx module
+---------------------------------------
+
+.. automodule:: xrpl.models.requests.account_tx
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.book\_offers module
+----------------------------------------
+
+.. automodule:: xrpl.models.requests.book_offers
+   :members:
+   :undoc-members:
+   :show-inheritance:
 
 xrpl.models.requests.channel\_authorize module
 ----------------------------------------------
@@ -29,10 +84,26 @@ xrpl.models.requests.channel\_verify module
    :undoc-members:
    :show-inheritance:
 
+xrpl.models.requests.deposit\_authorized module
+-----------------------------------------------
+
+.. automodule:: xrpl.models.requests.deposit_authorized
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 xrpl.models.requests.fee module
 -------------------------------
 
 .. automodule:: xrpl.models.requests.fee
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.gateway\_balances module
+---------------------------------------------
+
+.. automodule:: xrpl.models.requests.gateway_balances
    :members:
    :undoc-members:
    :show-inheritance:
@@ -85,6 +156,22 @@ xrpl.models.requests.manifest module
    :undoc-members:
    :show-inheritance:
 
+xrpl.models.requests.no\_ripple\_check module
+---------------------------------------------
+
+.. automodule:: xrpl.models.requests.no_ripple_check
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.path\_find module
+--------------------------------------
+
+.. automodule:: xrpl.models.requests.path_find
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 xrpl.models.requests.ping module
 --------------------------------
 
@@ -109,6 +196,14 @@ xrpl.models.requests.request module
    :undoc-members:
    :show-inheritance:
 
+xrpl.models.requests.ripple\_path\_find module
+----------------------------------------------
+
+.. automodule:: xrpl.models.requests.ripple_path_find
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 xrpl.models.requests.server\_info module
 ----------------------------------------
 
@@ -125,10 +220,74 @@ xrpl.models.requests.server\_state module
    :undoc-members:
    :show-inheritance:
 
+xrpl.models.requests.sign module
+--------------------------------
+
+.. automodule:: xrpl.models.requests.sign
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.sign\_and\_submit module
+---------------------------------------------
+
+.. automodule:: xrpl.models.requests.sign_and_submit
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.sign\_for module
+-------------------------------------
+
+.. automodule:: xrpl.models.requests.sign_for
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.submit module
+----------------------------------
+
+.. automodule:: xrpl.models.requests.submit
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.submit\_multisigned module
+-----------------------------------------------
+
+.. automodule:: xrpl.models.requests.submit_multisigned
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.submit\_only module
+----------------------------------------
+
+.. automodule:: xrpl.models.requests.submit_only
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
 xrpl.models.requests.subscribe module
 -------------------------------------
 
 .. automodule:: xrpl.models.requests.subscribe
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.transaction\_entry module
+----------------------------------------------
+
+.. automodule:: xrpl.models.requests.transaction_entry
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+xrpl.models.requests.tx module
+------------------------------
+
+.. automodule:: xrpl.models.requests.tx
    :members:
    :undoc-members:
    :show-inheritance:
