--- docs/source/xrpl.models.requests.transactions.rst
@@ -1,77 +1,0 @@
-xrpl.models.requests.transactions package
-=========================================
-
-Submodules
-----------
-
-xrpl.models.requests.transactions.sign module
----------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.sign
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.sign\_and\_submit module
-----------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.sign_and_submit
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.sign\_for module
---------------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.sign_for
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.submit module
------------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.submit
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.submit\_multisigned module
-------------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.submit_multisigned
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.submit\_only module
------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.submit_only
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.transaction\_entry module
------------------------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.transaction_entry
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-xrpl.models.requests.transactions.tx module
--------------------------------------------
-
-.. automodule:: xrpl.models.requests.transactions.tx
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: xrpl.models.requests.transactions
-   :members:
-   :undoc-members:
-   :show-inheritance:
