+++ tests/models/requests/test_sign.py
@@ -1,7 +1,8 @@
 from unittest import TestCase
 
+from xrpl import CryptoAlgorithm
 from xrpl.models.exceptions import XRPLModelException
-from xrpl.models.requests.transactions.sign import Sign
+from xrpl.models.requests import Sign
 from xrpl.models.transactions import AccountSet
 
 _ACCOUNT = "r9LqNeG6qHxjeUocjvVki2XR35weJ9mZgQ"
@@ -67,5 +68,7 @@ class TestSign(TestCase):
         self.assertTrue(request.is_valid())
 
     def test_valid_seed(self):
-        request = Sign(transaction=_TRANSACTION, seed=_SEED, key_type="secp256k1")
+        request = Sign(
+            transaction=_TRANSACTION, seed=_SEED, key_type=CryptoAlgorithm.SECP256K1
+        )
         self.assertTrue(request.is_valid())
