+++ xrpl/account/main.py
@@ -3,7 +3,7 @@
 from typing import Any, Dict, cast
 
 from xrpl.clients import Client
-from xrpl.models.requests.accounts.account_info import AccountInfo
+from xrpl.models.requests.account_info import AccountInfo
 from xrpl.models.requests.fee import Fee
 from xrpl.models.response import Response
