+++ xrpl/constants.py
@@ -5,8 +5,8 @@ from enum import Enum
 class CryptoAlgorithm(str, Enum):
     """Represents the supported cryptography algorithms."""
 
-    ED25519 = "ED25519"
-    SECP256K1 = "SECP256K1"
+    ED25519 = "ed25519"
+    SECP256K1 = "secp256k1"
 
 
 class XRPLException(Exception):
