+++ xrpl/keypairs/main.py
@@ -1,6 +1,6 @@
 """Public interface for keypairs module."""
 from secrets import token_bytes
-from typing import Dict, Optional, Tuple, Type, cast
+from typing import Dict, Optional, Tuple, Type
 
 from typing_extensions import Final
 
@@ -61,10 +61,8 @@ def derive_keypair(seed: str, validator: bool = False) -> Tuple[str, str]:
     decoded_seed, algorithm = addresscodec.decode_seed(seed)
     module = _ALGORITHM_TO_MODULE_MAP[algorithm]
     public_key, private_key = module.derive_keypair(decoded_seed, validator)
-    signature = module.sign(cast(str, _VERIFICATION_MESSAGE), private_key)
-    if not module.is_valid_message(
-        cast(str, _VERIFICATION_MESSAGE), signature, public_key
-    ):
+    signature = module.sign(_VERIFICATION_MESSAGE, private_key)
+    if not module.is_valid_message(_VERIFICATION_MESSAGE, signature, public_key):
         raise XRPLKeypairsException(
             "Derived keypair did not generate verifiable signature",
         )
@@ -102,7 +100,7 @@ def sign(message: bytes, private_key: str) -> str:
     return (
         _get_module_from_key(private_key)
         .sign(
-            cast(str, message),
+            message,
             private_key,
         )
         .hex()
@@ -124,7 +122,7 @@ def is_valid_message(message: bytes, signature: bytes, public_key: str) -> bool:
         True if message is valid given signature and public key.
     """
     return _get_module_from_key(public_key).is_valid_message(
-        cast(str, message),
+        message,
         signature,
         public_key,
     )
