+++ xrpl/models/__init__.py
@@ -1,1 +1,4 @@
 """Top-level exports for the models package."""
+from xrpl.models.exceptions import XRPLModelException
+
+__all__ = ["XRPLModelException"]
