+++ xrpl/models/requests/account_currencies.py
@@ -32,4 +32,4 @@ class AccountCurrencies(Request):
     ledger_hash: Optional[str] = None
     ledger_index: Optional[Union[str, int]] = None
     method: RequestMethod = field(default=RequestMethod.ACCOUNT_CURRENCIES, init=False)
-    strict: Optional[bool] = False
+    strict: bool = False
