--- xrpl/models/requests/accounts/__init__.py
@@ -1,30 +1,0 @@
-"""Account request models."""
-from xrpl.models.requests.accounts.account_channels import AccountChannels
-from xrpl.models.requests.accounts.account_currencies import AccountCurrencies
-from xrpl.models.requests.accounts.account_info import AccountInfo
-from xrpl.models.requests.accounts.account_lines import AccountLines
-from xrpl.models.requests.accounts.account_objects import (
-    AccountObjects,
-    AccountObjectType,
-)
-from xrpl.models.requests.accounts.account_offers import AccountOffers
-from xrpl.models.requests.accounts.account_tx import AccountTx
-from xrpl.models.requests.accounts.gateway_balances import GatewayBalances
-from xrpl.models.requests.accounts.no_ripple_check import (
-    NoRippleCheck,
-    NoRippleCheckRole,
-)
-
-__all__ = [
-    "AccountChannels",
-    "AccountCurrencies",
-    "AccountInfo",
-    "AccountLines",
-    "AccountObjects",
-    "AccountObjectType",
-    "AccountOffers",
-    "AccountTx",
-    "GatewayBalances",
-    "NoRippleCheck",
-    "NoRippleCheckRole",
-]
