+++ xrpl/models/requests/channel_authorize.py
@@ -1,12 +1,15 @@
 """
 The channel_authorize method creates a signature that can
 be used to redeem a specific amount of XRP from a payment channel.
+
+`See channel_authorize <https://xrpl.org/channel_authorize.html>`_
 """
 from __future__ import annotations
 
 from dataclasses import dataclass, field
 from typing import Dict, Optional
 
+from xrpl import CryptoAlgorithm
 from xrpl.models.requests.request import Request, RequestMethod
 from xrpl.models.required import REQUIRED
 from xrpl.models.utils import require_kwargs_on_init
@@ -18,6 +21,8 @@ class ChannelAuthorize(Request):
     """
     The channel_authorize method creates a signature that can
     be used to redeem a specific amount of XRP from a payment channel.
+
+    `See channel_authorize <https://xrpl.org/channel_authorize.html>`_
     """
 
     method: RequestMethod = field(default=RequestMethod.CHANNEL_AUTHORIZE, init=False)
@@ -27,7 +32,7 @@ class ChannelAuthorize(Request):
     seed: Optional[str] = None
     seed_hex: Optional[str] = None
     passphrase: Optional[str] = None
-    key_type: Optional[str] = None
+    key_type: Optional[CryptoAlgorithm] = None
 
     def _get_errors(self: ChannelAuthorize) -> Dict[str, str]:
         errors = super()._get_errors()
