+++ xrpl/models/transactions/__init__.py
@@ -17,7 +17,7 @@ from xrpl.models.transactions.payment_channel_create import PaymentChannelCreate
 from xrpl.models.transactions.payment_channel_fund import PaymentChannelFund
 from xrpl.models.transactions.set_regular_key import SetRegularKey
 from xrpl.models.transactions.signer_list_set import SignerListSet
-from xrpl.models.transactions.transaction import Transaction
+from xrpl.models.transactions.transaction import Memo, Signer
 from xrpl.models.transactions.trustset import TrustSet
 
 __all__ = [
@@ -30,6 +30,7 @@ __all__ = [
     "EscrowCancel",
     "EscrowCreate",
     "EscrowFinish",
+    "Memo",
     "OfferCancel",
     "OfferCreate",
     "Payment",
@@ -38,7 +39,7 @@ __all__ = [
     "PaymentChannelCreate",
     "PaymentChannelFund",
     "SetRegularKey",
+    "Signer",
     "SignerListSet",
-    "Transaction",
     "TrustSet",
 ]
