+++ xrpl/models/transactions/account_set.py
@@ -10,16 +10,18 @@ from __future__ import annotations  # Requires Python 3.7+
 from dataclasses import dataclass, field
 from typing import Dict, Optional
 
+from typing_extensions import Final
+
 from xrpl.models.transactions.transaction import Transaction, TransactionType
 from xrpl.models.utils import require_kwargs_on_init
 
-_MAX_TRANSFER_RATE = 2000000000
-_MIN_TRANSFER_RATE = 1000000000
-_SPECIAL_CASE_TRANFER_RATE = 0
+_MAX_TRANSFER_RATE: Final[int] = 2000000000
+_MIN_TRANSFER_RATE: Final[int] = 1000000000
+_SPECIAL_CASE_TRANFER_RATE: Final[int] = 0
 
-_MIN_TICK_SIZE = 3
-_MAX_TICK_SIZE = 15
-_DISABLE_TICK_SIZE = 0
+_MIN_TICK_SIZE: Final[int] = 3
+_MAX_TICK_SIZE: Final[int] = 15
+_DISABLE_TICK_SIZE: Final[int] = 0
 
 
 @require_kwargs_on_init
